[#6176] Resolved unexpected std::bad_alloc in fixed_buffer_resource unit test.
With the use of Boost v1.78, the fixed_buffer_resource class and unit
test were enabled. This revealed that the unit test was failing. The
failures were due to memory alignment fixes in the
fixed_buffer_resource.

The previous version of the unit test did not consider padding bytes
caused by aligning memory.

This commit resolves the issue by increasing the size of the buffer used
in the unit test.