[#5479] Load client API plugins in client API tests
Some unit tests require a running server so that they can test client
APIs through a client connection. The authentication in 4.3.0 requires
that client API plugins be loaded so that the new authentication plugin
framework can be used. Calls to do so have been added to all affected
unit tests.