[#5479] Auth plugin framework docs and enhancement
Adds documentation to existing authentication plugin framework class and
free functions as well as improving usability and some formatting
issues.

Adds a new function to the authentication plugin framework header which
allows a client or server to determine which authentication method it
ought to be using in a unified fashion. The implementation at this time
is, of course, a simple version check, but it could be based on other
criteria which can be queried from an RcComm.

The original auth plugin framework implementation includes some common
functions which are now available to any plugin implementing the
framework. These include a request message verification function and a
convenience function for invoking the client API endpoint for performing
an authentication plugin operation.