[#5479] Add legacy PAM authentication plugin for new framework
This ports the legacy PAM authentication plugin to the new
authentication plugin framework. It performs a simple password check
using the PAM API in a separately built application which is called from
the plugin on the server. Then, the native authentication flow is
employed to complete authentication with the iRODS server.