[#3727] Fix auth objects resolving/loading plugins
In the event that an auth object fails to resolve an auth plugin
interface, it may be the case that the plugin hasn't been loaded yet.
The failure should indicate that an attempt to load the plugin should be
made before returning an error. On success, a resolved plugin interface
will be returned and the resolution will be considered successful.