[#6323] Dequeue delay rules not found in catalog
When a rule in the queue is not found in the catalog for whatever
reason, it should be removed from the delay queue without doing any
additional work. This can happen due to a timing issue wherein the delay
server fetches work from the catalog that is currently being done and
then place it back on the queue after it has been removed. This is
normal and a side effect of the asynchronous producer-consumers model
employed by the delay server.