[#6357] Replace start with restart in control plane test
test_control_plane.TestControlPlane.test_shutdown_with_invalid_host
attempts to create a resource with an invalid host name and then use
irods-grid to shutdown the server. At the end of the test, the server is
started again and the resource is cleaned up.

There is an issue where the delay server migration logic keeps the main
server process alive after it stops receiving connections and this was
causing the start call to fail at the end of the test because the main
server process still exists. Until a cleaner solution is implemented, we
will use restart instead to ensure that the server comes back up cleanly.