[#6382] Remove supplied password support from PAM
This change removes support for a password supplied via keyword to the
plugin. The password should only be allowed through the prompt given by
the plugin.

This change adds support for a new keyword in the authentication plugins
called force_password_prompt. As the name suggests, this forces the
plugin to prompt the user for a password rather than using any password
supplied by the user beforehand or in an authentication file. This is
useful for clients like iinit which want to re-prompt the user for their
password even when they are currently authenticated.