[#5479] Skip calling server-side auth operations
The server-side authentication operations are not implemented or
supported in the new authentication plugin framework. This is partially
because the most-used/supported plugins do not implement these
operations and the ones that do implement the operations can probably
invoke the server-side functionality as part of the client
authentication flow. For now, we will not be compiling these bits as
they will be unused in 4.3.0.