[#6414] test_ssl: Increase key size to 2048 from 1024
The key size used to generate the RSA key for SSL communications in the
test has been 1024 for a long time in order to speed up the time taken
for each test. In newer versions of libssl, the key size requirements
have shifted such that 1024 is too small. This change updates the size
of the keys used in our SSL tests from 1024 to 2048.