[#6409] Check before inserting object metamap entry
When inserting an object metamap entry (that is, an entry into the table
which tracks associations of iRODS objects and metadata AVU IDs), we
need to make sure that the association does not already exist in the
object metamap. In the event that it already exists, the metamap entry
should not be inserted because this will raise an error.