[#6409][#6416] Expand accepted error codes in imeta addw test
In versions of unixODBC newer than 2.3.4, different error codes are
retured when attempting to insert multiple entries into the database
with the same values, specifically with R_OBJT_METAMAP. This difference
in error codes led to the discovery of a bug in our implementation and
the conclusion that addw should be deprecated and removed. This change
updates the addw test which exercises this behavior to accept either
error code.