[#6425] Fix next_executor() in delay server
std::atomic<int> will not protect against race conditions when
determining the index of the next executor in the list of configured
remote executors for the delay server. Instead, we should use a
std::scoped_lock with a std::mutex around a plain old int.

There is also a fault in the logic when incrementing the index to rotate
through the list of executors.

The result of the expression used to determine the index will always
be 0. When assigning the new value of index, the post-increment is the
last expression evaluated before the assignment occurs, and its return
value is the value of the operator before incrementing. The increment
has moved such that it is appropriately set for the next invocation.