[#6429] Fixed shutdown process for agent factory.
The signal handler for the agent factory uses exit(1) to terminate. This
works as long as the main iRODS process did not respawn a new agent
factory.

However, if the main iRODS server process had to fork a new agent
factory, the use of exit(1) can corrupt the parent's memory, breaking
the shutdown process. This results in the agent factory becoming an
orphan.

This commit fixes that issue by replacing the use of exit(1) with
_exit(1), as recommended by the POSIX standard.