[#6504] Make AVU-by-condInput association atomic
When the METADATA_INCLUDED_KW is used with some operations (e.g. put),
the AVU triples are said to be applied atomically. This change uses the
atomic metadata operations API to ensure that the association of the
metadata AVUs to the object is actually atomic as it relates to the
insertion of the AVUs and the association with the data object.

Note that this change does NOT make the AVU association atomic as it
relates to the creation of the data object in the catalog.

Additionally, this resolves an issue where the presence of an existing
metadata AVU-object association would result in an error on some
platforms due to the ODBC implementation. The atomic metadata operations
API checks for the existence of the AVU-object association before
inserting which allows us to sidestep the issue.