[#6597] Specify OpenSSL versions
On newer platforms, some of the functions and types in use by the SSL
networking plugin have been deprecated or hidden so we need to set a
compiler definition to indicate the version of SSL we would like to be
compatible with until we update these calls.

Specifically, OpenSSL 3.0 deprecated many low-level APIs and types in
favor of a new kind of interface.

The version of libssl in our declared dependencies needs to be changed
conditionally based on the version of OpenSSL detected by find_package.