[#6538] Provide mechanism for linking object libraries to unit tests
Plugins are built as module library targets. This tells CMake that they are not
meant to be installed or used in the same way as shared library targets. This
is what we want. However, this presents a hurdle when it comes to unit testing
code in our plugins, as module library targets cannot be linked against in
CMake.

What we must do instead is split out the code that is to be tested into an
object library target, and link the unit test against that instead. This change
seeks to make linking unit tests to object library targets trivially possible,
by adding a IRODS_TEST_LINK_OBJLIBRARIES CMake variable that can be defined
with a list of object library targets to be linked against.