[ 6622] Add test for closeAllL1Desc
Adds a test which opens a data object, executes some nonsense via
msiExecCmd, and then closes the data object. This demonstrates an issue
with msiExecCmd in which the open L3 descriptors are closed in the child
process forked to perform the msiExecCmd. This should only be happening
in the agent process.