[#6639] Refactored client_api_allowlist interface.
The original interface required that the developer call a static
function to get a handle to a singleton instance of the class before it
could be used. This makes usage more difficult and it doesn't offer any
benefits.

This commit improves the situation by changing the interface to a set of
namespaced free functions similar to the zone administration library.

This commit also makes the interface consistent with other modern iRODS
libraries.