[#6701] Refactored user administration library's interface.
This commit improves the interface by removing the need for developers
to check every API call. Instead, the developer can focus on the task at
hand. If an operation fails, the developer can catch it and react. This
improves readability and makes consumer code less noisy.

This commit also makes the interface consistent with other modern iRODS
libraries.