[#6693] test_auth: Newer OpenSSL requires 2048 numbits
Two tests in test_auth were failing to complete SSL handshakes with the
server when run on Ubuntu 20.04. This is because the command used to
generate the certificate chain file was using too low of a numbits
parameter. In OpenSSL 1.1.1, this apparently needs to be 2048 bits or
greater. In OpenSSL 1.1.0, 1024 was sufficiently large and so the test
was passing on Ubuntu 18.04.