[#6687] Delay server migration now uses host_resolution config.
Delay server migration does not cover the case where a server has
multiple names (i.e. FQDNs). This constrains administrators to using
the name that matches the output of gethostname().

This commit resolves the issue by adjusting the delay server migration
logic so that it consults the host_resolution property in
server_config.json.