[#6814] Address sanitizer is manually run for special cases.
_exit() does not invoke handlers registered via std::atexit() or
on_exit().

This is a problem for Clang's address sanitizer because it relies on
these facilities to report its findings.

To get around this issue, __lsan_do_leak_check() is invoked at locations
that guarantee findings are reported. This is especially necessary for
the delay server because it is launched via fork-exec. If the call to
exec() succeeds, the std::atexit/on_exit handler registrations are
removed.