[#6634] Made it so that agents can capture the client IP address.
The iRODS server originally ignored socket file descriptor values of
zero. If iRODS detected a socket file descriptor of zero, it would not
capture the client's IP address, breaking APIs that depend on it.

With recent efforts to clean up the server implementation, zero now
appears as a socket file descriptor. Therefore, zero must be allowed.