[#6634] Fixed segfault issue stemming from refactored clearMsParam().
clearMsParam()'s logic was adjusted to resolve memory leaks in the
server. However, that change caused resulted in the iRODS Rule Language
triggering a segfault (dereferencing a nullptr).

To avoid the segfault, the original implementation of clearMsParam() was
restored as a new function called clear_ms_param(). This new function is
private to the implementation of the iRODS Rule Language.