[#6428] Use service account for rsModColl redirect
When redirecting to the catalog service provider in rsModColl, users
which have been temporarily granted privileged access in the agent
connection will lose the privileged status on the server-to-server
connection made there.

This change establishes a temporary connection as the rodsadmin
authenticated in the local service account so that the privileges remain
in tact when modifying the collection. This is important so that
collection mtimes can be updated after an operation is completed even if
the user does not have any permissions on the parent collection.

NOTE: This may have performance implications, but is required for
updating collection mtimes to operate completely correctly.