[ 6997] Add test mechanism for plugin operation failure
This change adds a check to the plugin framework which checks plugin
property maps for the "munge_operations" keyword. Currently, the
keyword must contain a value that is a comma-delimited list of key-value
pairs separated by colons. This is meant to provide a means for testing
to configure a plugin (at this time, only resource plugins are being
tested) such that certain operations return specific error codes.

This adds some tests which ensure that the system is behaving properly
when certain resource plugin operations fail. Specifically, when a data
object is created and some resource plugin operation fails, the data
object should be marked stale to ensure that the state of the data is
properly reflected in the catalog.