[#6479] replica_close: Close replica before finalize
Some resource plugins cannot stat before l3Close has been invoked. This
change closes the physical file descriptor (i.e. calls l3Close) before
attempting to finalize catalog information because stats on the physical
data occur during this process.

This change ensures that the replica status is only updated when
update_status option is true. If it is false, the user has explicitly
disabled updating the status, so that should be honored.

This change also ensures that the L3 descriptor index is set to an
invalid value after physically closing the object so that replica_close
can be called multiple times, should that be required by the client
application.