[#6378] Always clean up and free rsComm members
Ensures that any data objects with open L1 descriptors are finalized,
including in situations where the network stop plugin operation fails or
physically closing the replica fails.

This change also makes sure that the put API cleans up after itself when
it fails to close the physical object on completion. This situation is
not uncommon with the S3 plugin where the actual upload to the storage
backend is not initiated until it is closed, and is the most likely time
for something to go wrong.