[#6896] Requested replica always wins the vote
When resolving a resource hierarchy with a "requested replica", the
requested replica should always win the vote even if it is not the
highest vote value. If the vote value is 0, this is an error and the
voting logic will return an error.

Even though this is providing an option for users to circumvent policy
configured in the server, administrators can still configure their
policy such that it is always honored even when the user attempts to
access a replica which would not be accessed by default due to the
values resulting from a vote. This has been the historical behavior for
the APIs which allow for using REPL_NUM_KW to target a specific replica
by number.

This change also adds a new free function to the voting library which
acts as a safe way to compare votes to 0. This is to avoid issues
associated with comparing floating point numbers.

This change also includes a peculiar workaround which is required in
order for voting to behave in the compound resource when it is
configured to prefer the archive resource (i.e. configured to "always"
overpave the cache replica). The replica in the cache resource must
contain a vote value if it is requested or it will cause operations
which request that replica by number to fail. Since the cache resource
only votes when the archive votes 0, this can happen frequently. As
such, the replica in the cache resource is now set to vote with the same
value as the replica in the archive resource (in the normal case)
because the replica in the archive resource is going to be overpaving
the replica in the cache resource anyway.