[#6900] Restored specific query, DataObjInCollReCur, for MySQL 8.
The specific query, DataObjInCollReCur, is not new. It was added years
ago as an optimization, but had to be removed because MySQL (prior to
version 8) did not support the SQL WITH-clause.

iRODS 4.3.1 does not support MySQL versions prior to 8.0 so it is safe
to restore the specific query.

The specific query will be added if the alias does not exist.