[#6896] Fix operation for vote in open API
The open API resolves the resource hierarchy if needed. It checks for
the presence of the O_CREAT flag in the openFlags and, if present,
resolves the hierarchy for the CREATE operation. Else, the hierarchy is
resolved for the OPEN operation. The OPEN operation is the operation for
open-for-read. This means that opens-for-write have been resolving
hierarchies as open-for-read operations, which is incorrect.

This change adds an additional check to see if a write flag is being
used when determining the operation. If present, the hierarchy is
resolved for the WRITE operation. If neither O_CREAT nor a write flag is
being used, the hierarchy is resolved for open-for-read.