[#6931] Use clearDataObjInfo for reg data obj output struct cleanup
While there are a few key differences between freeDataObjInfo - which is
the function which was previously used as the clearOutStruct for the
register data object API - and clearDataObjInfo, the way the
clearOutStruct mechanism is being used in the API handler requires that
any out struct pointers which are freed be set to nullptr. This is
because the API handler calls std::free on the out struct pointer after
calling the clearOutStruct function. This could result in a double free.
Instead, the dataObjInfo struct itself should be left unfreed while all
of its members are freed as part of cleaning up the out struct.