[#6502] Updated the stage-to-cache comparison check on write
For write operation, the voting of stale is higher than good.
For the purposes of determining whether stage-to-cache is necessary,
a stale cache and good archive should trigger the staging-to-cache.
A stale archive and good cache should not.  To get this behavior
the vote comparison between cache/archive has to be flipped for writes.