[#6956] Config reload test asserts changes in test-running host
This change fixes test_reload_command__issue_6529 by adjusting the
assertions being made on the output of irods-grid reload --all. This
command will return a set of changes for all servers in the zone because
the reCacheSalt configuration is always captured whenever the
configuration is loaded. As a result, all servers in the zone will
return a set of changes. This test only cares about the changes
made to the configuration on the local server.