[#7000] Fix tests for running in topology
Test_AllRules: Use zonereport to detect database

Test_AllRules skips one test based on the database type. This is derived
from the local server's server_config.json file. Catalog consumers do
not have the database configuration in their server_config.json files,
so this fails when running topology tests on catalog consumers.

This change uses izonereport to find the database type in the
catalog provider entry of the zonereport. This should work no matter which
server in the zone is running the tests.

---

msiExec PEP test should use local resource

test_rulebase.Test_Rulebase.test_msiExecCmd_closeAllL1Desc__issue_6623
implements a resource PEP but the data object being opened only has a
replica which is on a resource tied to a different computer, so the PEP
fires over there (and is, of course, not implemented anywhere but the
local machine).

This change makes a local resource and puts the data object's only
replica there so that the resource PEP will fire.

---

Avoid log test in test_acPreProcForExecCmd__3867

Use metadata on a temporary resource instead because the logs are
unreliable.

---

Skip msiDeleteUnusedAVUs for consumer topology tests

msiDeleteUnusedAVUs does not redirect to the catalog provider, so this
microservice will always fail when not run from the catalog provider.

---

Fix hostname in downgrade service account admin test

The test for downgrading the service account rodsadmin account needs
to determine the expected hostname by fetching it from the zone report
just like rsGeneralAdmin. Apparently, the order of the entries is not
consistent across test runs but is consistent in a given server's
lifetime.

---

Fix recursive ireg tests expecting success

Two recursive ireg tests were expecting success when registering files
to a resource on a machine on which those files do not exist. A local
resource is now created and targeted for registration.

---

Skip test_pam_password_authentication in topology

Configuring SSL is part of the test in order to ensure that the test
is run regularly for normal testing. This cannot be done for other
servers in the topology from inside the test (must be done by the
testing environment before running the test). Skip the test if run
in topology since the topology cannot be configured properly to run it.

---

Change expected error in msiRemoveUserFromGroup test

If the test is being run from a catalog consumer, the privilege level
of the authenticated user is checked before redirecting to the catalog
provider and returns an error. When run on the catalog provider, the
database operation itself will return the error.