[#6976] Fix tests for running in topology
Test_AllRules: Use zonereport to detect database

Test_AllRules skips one test based on the database type. This is derived
from the local server's server_config.json file. Catalog consumers do
not have the database configuration in their server_config.json files,
so this fails when running topology tests on catalog consumers.

This change uses izonereport to find the database type in the
icat_server entry of the zonereport. This should work no matter which
server in the zone is running the tests.

---

Restart server before tests reading logs

The log stops receiving messages sometimes and this is alleviated by
restarting the server. Rather than rewriting all of these tests, let's
just restart the server for now so that the log message checks will work
more consistently.

---

msiExec PEP test should use local resource

test_rulebase.Test_Rulebase.test_msiExecCmd_closeAllL1Desc__issue_6623
implements a resource PEP but the data object being opened only has a
replica which is on a resource tied to a different computer, so the PEP
fires over there (and is, of course, not implemented anywhere but the
local machine).

This change makes a local resource and puts the data object's only
replica there so that the resource PEP will fire.

---

Avoid log test in test_acPreProcForExecCmd__3867

Use metadata on a temporary resource instead because the logs are
unreliable.

---

Skip msiDeleteUnusedAVUs for consumer topology tests

msiDeleteUnusedAVUs does not redirect to the catalog provider, so this
microservice will always fail when not run from the catalog provider.

---

Fix hostname in downgrade service account admin test

The service account rodsadmin is the same across the zone for testing in
topology. The error message displayed when attempting to downgrade the
privileges of the service account rodsadmin will display the hostname of
the catalog service provider, which will not be the hostname of the
local server when running tests on a catalog consumer.