[#6997] Fix stage-to-cache with read-only permission
The compound resource must stage an object to cache before it can be
read. If a replica exists only in the archive resource, users with
read-only permissions on the data object who attempt to read it will
trigger a stage-to-cache operation. This naturally requires that the
user be allowed to write a new replica for the data object in the cache
resource. This is not allowed because the user only has read
permissions.

This change temporarily grants object ownership to the user when the
stage-to-cache operation occurs.