[#6428] Add test for irm of object in read-only remote collection
This change adds a test in which a user has read permission on a remote
collection and own permission on an object inside the collection. The
user attempts to remove the object from the collection and ensures that
the collection mtime is updated appropriately while no errors occur on
the user's end.