[#6997] Stage-to-cache is performed by admin user
This change temporarily elevates privileges for the client user when
performing stage-to-cache replication in the compound resource. The
client user is also switched to be the local service account rodsadmin
so that the elevated privileges will hold through redirections.

Staging to cache in the compound resource has been updated such that all
existing server-to-server connections are disconnected after assuming
the admin client identity. This ensures that any redirects establish a
new server-to-server connection using the appropriate client identity.
After staging to cache has completed, all existing server-to-server
connections are disconnected again to ensure that the original client
identity is used again in any future redirects.