[#6997] Add admin keyword support to trim_replica
This change adds an admin option to the trim_replica function so that the
admin keyword will be added to the input for the trim API. The option
defaults to false to maintain the existing behavior, but the compound
resource needs this to remove cache replicas which were created as a
result of a failed stage-to-cache due to the data not yet having been
retrieved from Glacier.

Also renames purge_cache to trim_replica.