[#7192] Session signature is now stored in RcComm.
This is required for clients managing multiple connections to function
properly.

Before this change, the implementation of the session signature was
designed for clients to manage a single connection. This was
accomplished through the use of a private global variable. Applications
using multiple connections meant the session signature always reflected
the most recent update to the private global variable, breaking the
ability for all but one of the connections to be usable in native
authentication password management.

By attaching the session signature to the RcComm, all connections gain
access to the correct session signature restoring the ability to change
passwords for native authentiation.