[#7193] Relaxed dependency on obfGetPw() in user administration library.
Changing the password of a user requires the caller's password. Before
this commit, the user administration library had a hard requirement on
obfGetPw(). This made it difficult to impossible to support use-cases
where there's no .irodsA file.

This commit resolves that by allowing developers to pass the caller's
password directly through the library, bypassing the need for the
.irodsA file.

This commit also replaces the use of obfEncodeByKeyV2() with
obfEncodeByKey(). This aligns the library's implementation with that of
iadmin and removes the need of the client-side session signature which
is important since the server side does not have anything equivalent to
the client-side session signature.