[#7099] Implemented new API endpoint: rx_check_auth_credentials
This API endpoint is a new feature for native authentication only.

It was designed to improve performance for client applications acting as
servers (e.g. the iRODS HTTP API).

The API endpoint takes user credentials for native authentication and
verifies if they are correct. No state within the client or server is
modified.