[#7250] GenQuery ends database transaction following ticket stat update.
If a ticket is enabled on a connection, running GenQuery would result in
an update to R_TICKET_MAIN. The update would be executed within a
transaction, but the transaction would never complete because there was
no call to commit or rollback afterwards.

This change fixes that by calling commit or rollback in the GenQuery
implementation rather than the database plugin.