[#7250] Added comments to help future readers.
This commit attempts to explain why two low level database functions are
implemented the way they are.

Specifically, the comments make it clear that the functions in this
commit ARE NOT suppose to COMMIT or ROLLBACK any database updates.
Instead, the functions are meant to be used within an existing
transaction.

The caller of these functions is responsible for completing the
transaction.