[#7274] Add pam_password configs to database on upgrade
This change adds a new namespace to the R_GRID_CONFIGURATION table in
the database for pam_password. The configuration values which were set
in the server_config plugin_configuration/authentication/pam_password
stanza are now to be set here. These can be configured with the iadmin
set_pam_password_config subcommand and queried using iadmin
get_pam_password_config.