[#7283] set_grid_configuration: Add protected config list
Adds a list of grid configuration namespaces and option names which are
not allowed to be modified via set_grid_configuration API. Any namespace
which has an option name of empty string means that the entire namespace
is protected.

This change also adds a check for a given namespace/option_name
modification attempt to return an appropriate error if it is not found.
This is to notify the administrator that the namespace/option_name they
attempted to update does not exist (whether this is due to a typo, an
update, or otherwise).