[#7303] Added new column to R_RESC_MAIN: modify_ts_millis
This new column stores the fractional part of modify_ts. The reason for
introducing the column is so the connection pool and other facilities
watching resources, have a way to detect changes which occur within the
same second.

modify_ts_millis will have the following properties:

    - A data type of varchar(3)
    - A default value of "000"

This commit also introduces a new GenQuery1 column which maps to
modify_ts_millis:

    - RESC_MODIFY_TIME_MILLIS