[#7274] Use grid configurations for auth configs
Authentication configurations in the database now use grid configurations
instead of server_config. These are stored in the database and so incur
an additional query whenever they are needed. If the values are invalid,
an error is returned, which renders all pam_password and native authentication
which uses TTL to be unusable. The values are updated on each invocation
of the database operations to update the temporary passwords.