[#7327] Filesystem library now understands 4.2 permission strings.
iRODS 4.3.0 changed the naming of permission strings. Each space in a
permission string was replaced with an underscore. The downside to this
change was the iRODS Filesystem library lost the ability to convert the
old permission strings to enumerations.

This commit resolves that by adding conversions for the following
permission strings:

    - "read object"
    - "modify object"

This commit also changes the conversion logic to throw an exception when
it cannot convert the passed string to a known permission enumeration.
This is important for identifying compatibility issues and holes in the
implementation.