[#7350] Connection pool properly checks size parameter before use.
Before this commit, it was possible to pass a negative integer value
through the connection pool's constructor to an internal std::vector.
This could possibly result in the std::vector allocating a large amount
of memory if its size parameter uses an unsigned integer type.

This commit addresses that issue by default initializing the
std::vector and checking the size parameter before using it to resize
the std::vector.