[#7476] Prefer good replicas for write votes
This commit changes the "preferred" replica status for votes on write
operations from stale to good.

One consequence of this decision is allowing replication operations to
to allow targeting good replicas for update. Instead of actually overwriting
the data in the good replica, fileModified is triggered directly to invoke
any policy defined by coordinating resources.

Clients can now request good replicas to be overwritten provided that the
source replica is good. Clients can now request that a replica overwrite
itself provided the source and destination replicas are good. In both cases,
no data movement occurs.

When the source and destination replicas are the same replica, the
replication operation is a no-op, although fileModified will be
triggered so that any configured policy will be in effect. If the
source replica is stale, it cannot be used to update any other
replicas regardless.