[#6954] Add clear_shm function and use in run_tests.py
Sometimes, the server dies and leaves behind shmem files which prevents
new shmem files from being created. This is a problem when running tests
in succession because when one test fails due to this unusual issue, the
rest of the tests in the set fail (more or less).

This adds a function which deletes all the irods_* files out of all of the
possible shmem locations. scripts/run_tests.py has been modified to stop
calling IrodsController.restart() before running tests and instead stops
the server, clears the shmem files if they are there, and starts the server
again.

Again, this function is only being used in run_tests.py as of this commit.