[#7498] Deprecate AGE_KW usage in trim API
Adds a deprecation message to the rError stack of the RsComm
object in the trim API when the AGE_KW is provided by the client.

This commit fixes an issue where the deprecation messages are not being
printed by itrim due to a client call occurring after rcDataObjTrim. This
causes the rError stack to be freed and set to nullptr in the RcComm which
means that the deprecation messages are lost when rcObjStat is called. The
issue would only manifest when a trim actually occurs such that the
rcObjStat call would be necessary. At the expense of a possibly unnecessary
API invocation, rcObjStat is now called before performing the trim so that
the messages from the salient API call (namely, rcDataObjTrim) can be
returned to the itrim client.

Also also, fixes a potential memory leak in trimUtil.