[#5359] irsync/iput: Add and update symlink tests
This commit adds a test for irsync -r --link with good and bad symlinks
and makes sure that irsync correctly ignores all of them.

This commit also adds a test for irsync -r with good symlinks and makes
sure that irsync correctly uploads all the data with the right names.

Equivalent tests were also added for iput.

Existing symlink tests have been updated to reflect the behavior of the
--link option in iput/irsync.