[#7556] Make find_replica a member function of data_object_proxy
This commit makes irods::experimental::data_object::find_replica a member
function of class template irods::experimental::data_object_proxy, removing
the free functions which required the caller to construct a data_object_proxy
in order to use it. Given that this is in the irods::experimental namespace,
the free functions have been removed.

This commit replaces the find_replica free function variants which have been
removed with a more useful set of functions which take a DataObjInfo&
instead of a data_object_proxy. The returned type is just a pointer to the
DataObjInfo in the list which matches the requested information for ease of
use with our existing internal interfaces.