[#6440] univMSSInterface: Replace 'which' with 'type -P'
When the Python Rule Engine Plugin is installed, the bash profile is
re-evaluated in such a way that "which" is modified to pick up function
names in bash in addition to and even instead of things in $PATH. This
leads to problems in univMSSInterface which uses several common GNU utilities
as function names.

Go team.

Co-authored-by: Markus Kitsinger (SwooshyCueb) <root@swooshalicio.us>
Co-authored-by: Terrell Russell <terrellrussell@gmail.com>
Co-authored-by: Derek D <ddong@renci.org>