[#6441,#7389] Add sbin directory to PATH for tests
Some tests use executables which live in the sbin directory. This
directory, due to differences in our various supported plaforms, is
not always in the $PATH of the iRODS service account, which executes
the tests. We need to ensure that the executables in the sbin directory
are available while the tests are running.

Also, there is a call to restart the server but not in test_mode.