[#7243] test_auth: Error on missing irodsauthuser
This commit raises an EnvironmentError when irodsauthuser is not present
and the test_auth suite is run. This is better than a test failure because
it should indicate that the prerequisites for the test are not being met
rather than something being wrong with the code that is being tested.