[#7521] test_control_plane: Increase delay server sleep time
Increasing the delay server's sleep time decreases the chances that
it will wake up and query for work from the server while it is being
shut down. This makes for cleaner shut downs which should prevent
future tests run on the same host from failing due to failure to
clean up shmem files or something of that sort.