[#7444,#7465] Add tests for irepl at various permission levels
There is already a test for admin users being able to replicate with
insufficient permissions: test_irepl_data_object_with_no_permission__4479

This commit adds a test for a regular rodsuser attempting to replicate
a data object on which they have all the various permission levels
available. There is a set that does not allow the user to even see the
data object exists, a level that lets the user see it but not replicate,
and a level that lets the user replicate. Each permission level exists in
a distinct unittest.subTest.