[ 7444, 7465] Prevent replication with insufficient permissions
This adds a check before attempting a replication to ensure that the
user has permission to perform the operation. At present, modify_object
is required to perform replication. There are permission levels between
read_object and modify_object which may allow users to create new
replicas but they will not be finalized properly. Until this is
resolved or the required permission level for replication is changed,
we need to ensure that only users with modify_object or higher can
replicate a data object.