[#7390] Auth tests: Loosen error message assertions
Due to an issue with how error messages are handled with password
expirations, some tests need to be updated to accept more than one
answer. This is because while the error messages are not ideal, the
state of affairs is considered acceptable and we need tests to pass.
Skipping the tests is less ideal still. The issue with the error
messages will be resolved in #7344.