[#7619] irodsctl stop: Fix orphan agent cleanup
The way the server is being shut down was reworked in commit
1a5b0399141de3448390a2d875290c60f669342e.

There is a place where a list of processes descending from the main
server process is being created by using the filter() built-in command.
The original implementation was written for python2 because all of our
supported platforms used python2. In python2, filter() calls return the
filtered iterable. Modern platforms we support use python3. In python3,
filter() returns a filter object, not an iterable.

There is no benefit to delaying generating the filtered set, so this
commit generates the list immediately by wrapping the call to filter()
with set().