[#7626] ModDataObjMeta: Fix ADMIN_KW usage with ALL_KW
The ADMIN_KW appears to not work when used with the ALL_KW. This is
because getDataObjInfoIncSpecColl is called without passing along the
ADMIN_KW. If the admin does not have permissions on the data object
in question, CAT_NO_ACCESS_PERMISSION is returned regardless of whether
the ADMIN_KW was used.

This commit passes the ADMIN_KW along to getDataObjInfoIncSpecColl.