[#7687] Change log level severity of fallback log message: warn -> trace
Before this commit, the lack of a log category in the "/log_level"
stanza of server_config.json would cause a warning to be written to the
log file every time the missing log category was used. This would result
in the the log file being filled with warning messages about falling
back to the default log level of "info" for that log category.

This commit makes it so that the fallback message is logged as a "trace"
level message, which keeps the log file quiet, but the message available
only if the server/logger is configured to output more messages.